1.0.11

+++ lib/Snapshot.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @copyright Copyright (c) 2017 Robin Appelman <robin@icewind.nl>
  *
@@ -63,11 +64,11 @@ class Snapshot {
 	}
 
 	public function getMtime($file): int {
-		return filemtime($this->getFilePath($file));
+		return (int)filemtime($this->getFilePath($file));
 	}
 
 	public function getSize($file): int {
-		return filesize($this->getFilePath($file));
+		return (int)filesize($this->getFilePath($file));
 	}
 
 	public function getSnapshotDate(): DateTime {
